#2066: more instances of 'unsigned long' used where we want uintptr_t
git-svn-id: https://xpra.org/svn/Xpra/trunk@21151 3bb7dfac-3a0b-4e04-842a-767bc560f471

+++ src/xpra/codecs/csc_libyuv/colorspace_converter.pyx
@@ -72,7 +72,7 @@ cdef inline uintptr_t roundupl(uintptr_t n, uintptr_t m):
     return (n + m - 1) & ~(m - 1)
 
 cdef inline uintptr_t memalign_ptr(uintptr_t ptr):
-    return <uintptr_t> roundupl(<unsigned long> ptr, MEMALIGN_ALIGNMENT)
+    return <uintptr_t> roundupl(<uintptr_t> ptr, MEMALIGN_ALIGNMENT)
 
 
 def init_module():
